Introduce VariableLookupTraverser, assure uniqueness of completions
