+++ lib/theme_check/language_server.rb
@@ -18,6 +18,7 @@ require_relative "language_server/variable_lookup_finder/assignments_finder"
 require_relative "language_server/variable_lookup_finder/constants"
 require_relative "language_server/variable_lookup_finder/liquid_fixer"
 require_relative "language_server/variable_lookup_finder"
+require_relative "language_server/variable_lookup_traverser"
 require_relative "language_server/diagnostic"
 require_relative "language_server/diagnostics_manager"
 require_relative "language_server/diagnostics_engine"
